requirements.txt
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
Deprecated==1.2.13
environ-config==22.1.0
idna==3.4
packaging==21.3
psycopg2-binary==2.9.5
pyparsing==3.0.9
pyTelegramBotAPI==4.7.1
redis==4.3.4
requests==2.28.1
urllib3==1.26.12
wrapt==1.14.1
